+++ packages/stream_chat_flutter/lib/src/message_list_view.dart
@@ -638,10 +638,14 @@ class _MessageListViewState extends State<MessageListView> {
     );
 
     final backgroundColor = MessageListViewTheme.of(context).backgroundColor;
+    final backgroundImage = MessageListViewTheme.of(context).backgroundImage;
 
-    if (backgroundColor != null) {
-      return ColoredBox(
-        color: backgroundColor,
+    if (backgroundColor != null || backgroundImage != null) {
+      return Container(
+        decoration: BoxDecoration(
+          color: backgroundColor,
+          image: backgroundImage,
+        ),
         child: child,
       );
     }